+++ app/sections/Work/projectswrapper.module.css
@@ -1,5 +1,4 @@
 .container {
   flex-direction: column;
   align-items: center;
-  gap: 72px;
 }
