+++ app/work/page.tsx
@@ -2,10 +2,12 @@ import React, { Suspense } from 'react'
 import Intro from '../sections/Work/Intro'
 import Solutions from '../sections/Work/Solutions'
 import Companies from '../sections/Work/Companies'
-import Projects from '../sections/Work/Projects'
+import ProjectsWrapper from '../sections/Work/ProjectsWrapper'
 import { ProjectsSkeleton } from '../components/skeletons'
+import Projects from '../components/Projects/Projects'
+import { fetchProjectsTotal } from '../lib/data'
 
-export default function Work({
+export default async function Work({
   searchParams,
 }: {
   searchParams?: {
@@ -14,14 +16,18 @@ export default function Work({
   }
 }) {
   const query = searchParams?.query || ''
-  const currentOffset = Number(searchParams?.offset) || 0
+	const currentOffset = Number(searchParams?.offset) || 0
+	
+	const total = await fetchProjectsTotal()
 
   return (
     <div className="main_wrapper">
       <Intro />
-      <Suspense fallback={<ProjectsSkeleton />}>
-        <Projects query={query} offset={currentOffset} />
-      </Suspense>
+      <ProjectsWrapper>
+        <Suspense fallback={<ProjectsSkeleton />}>
+          <Projects offset={currentOffset} />
+        </Suspense>
+      </ProjectsWrapper>
       <Companies />
       <Solutions />
     </div>
